+++ e2e/node/basic/mainnet.test.ts
@@ -33,10 +33,9 @@ const createWhoamiActor = async (identity: Identity) => {
 describe('certified query', () => {
   it('should verify a query certificate', async () => {
     const actor = await createWhoamiActor(new AnonymousIdentity());
-
     const result = await actor.whoami();
     expect(Principal.from(result)).toBeInstanceOf(Principal);
-  }, 100_000);
+  }, 10_000);
   it('should verify lots of query certificates', async () => {
     let count = 1;
     const identities = Array.from({ length: 20 }).map(() => {