+++ docs/CHANGELOG.md
@@ -5,14 +5,16 @@
 ### Added
 
 - chore: adds required `npm audit` check to PRs
+- new `HttpAgent` option: `backoffStrategy` - allows you to set a custom delay strategy for retries. The default is a newly exported `exponentialBackoff`, but you can pass your own function to customize the delay between retries.
 
 ### Changed
 
 - chore: upgrades github actions to v4
+- fix: retry logic now includes delays with exponential backoff matching the dfx strategy. Retries should no longer happen too quickly for the replica to catch up.
 
 ## [1.2.1] - 2024-04-25
 
-### Changed
+### Added
 
 - feat: make `IdbStorage` `get/set` methods generic
 - chore: add context to errors thrown when failing to decode CBOR values.